Introduce 2-stage resampling to speed up slice rendering
In slicer3 the all slice views were resampled to the full window size
using vtkImageResliceMask, meaning that there was a 3D interpolation
being done where it really wasn't needed.  This branch tries to improve
performance by doing 3D reslicing only a the resolution of the input
data and then doing 2D zooming of the result.