Introduce 2-stage resampling in attempt to speed up slice rendering
In slicer3 the all slice views were resampled to the full window size
using vtkImageResliceMask, meaning that there was a 3D interpolation
being done where it really wasn't needed.  This branch tries to improve
performance by doing 3D reslicing only a the resolution of the input
data and then doing 2D zooming of the result.

Question remains if resampling the 2D image should be done in software
or as part of the rendering pipeline (probably the latter) and how
the image should be draw to optimize opengl performance on all platforms.
Options are texture based or pixel bitblt (vtkImageActor or vtkImageMapper)